feat: add GitHub Actions workflow for installing packages
Added a new GitHub Actions workflow named `Install` to facilitate the installation of various packages. The workflow includes inputs for GitHub Token, APT packages, Brew packages, Chocolatey packages, and other packages. The workflow runs composite steps to execute scripts for installing the specified packages based on the inputs provided.

The purpose of this change is to streamline the process of installing packages in different environments by leveraging GitHub Actions.