fix(common): disable isAdminEnforced for main branch and skip shellcheck in pre-commit config
In the sync-repo-settings.yaml file, the isAdminEnforced setting for the main branch is disabled to allow non-administrators to make changes. Additionally, in the pre-commit-config.yaml file, the shellcheck tool is skipped in the continuous integration process to improve efficiency.

These changes were made to enhance the flexibility and efficiency of the repository settings and pre-commit checks.