feat(common): add Renovate configuration file
Added a new Renovate configuration file `renovate.json` to the `.github` directory in the `common` folder. The file includes settings to extend recommended configurations and automerge all default configurations. Additionally, it specifies git ignored authors to exclude commits from the pre-commit-ci bot. This change aims to automate dependency updates and ensure the project stays up-to-date with the latest dependencies.